feat(array): allow negative positional indices field[-N] (#158)
The accepted SEP (#212) promotes negative indices to specified: [-1] is the
last element, [-k] the k-th from the end, with out-of-range in either
direction yielding no match (matching SPL mvindex(..., -1) and KQL arr[-1]).

The parser keeps a signed index in the literal field path; the evaluator
resolves it against the array length via a shared resolver; PostgreSQL JSONB
emits the native negative subscript (->-1, PG 11+). The convert capability
guard recognizes negative indices too, so backends that cannot lower them
still fail loudly.

+++ crates/rsigma-convert/src/backends/postgres/mod.rs
@@ -33,10 +33,11 @@ fn validate_sql_identifier(s: &str) -> Result<()> {
     }
 }
 
-/// A JSONB field-path navigation op: an object key or a positional index.
+/// A JSONB field-path navigation op: an object key or a positional index
+/// (possibly negative, counting from the end of the array).
 enum FieldOp {
     Key(String),
-    Index(u32),
+    Index(i64),
 }
 
 /// Return a fresh, query-unique sequence number for naming array-element
@@ -393,7 +394,7 @@ impl PostgresBackend {
                     let close = rest
                         .find(']')
                         .ok_or_else(|| ConvertError::InvalidIdentifier(part.to_string()))?;
-                    let n: u32 = rest[..close]
+                    let n: i64 = rest[..close]
                         .parse()
                         .map_err(|_| ConvertError::InvalidIdentifier(part.to_string()))?;
                     ops.push(FieldOp::Index(n));
